Add flag for APIs that require policy install
